리뷰 완료

+++ src/main/java/com/example/aboutme/user/UserResponseRecord/ExpertMainDTO/RecentReviewRecord.java
@@ -3,7 +3,7 @@ package com.example.aboutme.user.UserResponseRecord.ExpertMainDTO;
 public record RecentReviewRecord(
         Integer reviewId,
         String clientName,
-        Double score,
+        Integer score,
         String content
 ) {
 }
